feat: carousel extract as a single component
